[KAN-86] rename: 만료시간 타이머  이름 변경
